undo some dumbs

+++ src/main/java/page/info/EnemyInfoTable.java
@@ -66,7 +66,7 @@ public class EnemyInfoTable extends Page {
 		double mula = mulatk * e.de.multi(b) / 100;
 		main[1][3].setText("" + (int) (e.de.getHp() * mul));
 		main[1][7].setText("" + (int) Math.floor(e.de.getDrop() * b.t().getDropMulti()) / 100);
-		main[2][3].setText("" + (int) (e.de.allAtk() * mula * 30 / e.de.getItv()));
+		main[2][3].setText("" + e.de.allAtk() * mula * 30 / e.de.getItv());
 		int[][] atkData = e.de.rawAtkData();
 		for (int i = 0; i < atks.length; i++)
 			atks[i][1].setText("" + Math.round(atkData[i][0] * mula));
@@ -193,7 +193,7 @@ public class EnemyInfoTable extends Page {
 		main[1][4].setText(MainLocale.INFO, "HB");
 		main[1][5].setText("" + e.de.getHb());
 		main[1][6].setText(MainLocale.INFO, "drop");
-		main[1][7].setText("" + (int) Math.floor(e.de.getDrop() * b.t().getDropMulti()) / 100);
+		main[1][7].setText("" + Math.floor(e.de.getDrop() * b.t().getDropMulti()) / 100);
 		main[2][0].setText(MainLocale.INFO, "range");
 		main[2][1].setText("" + e.de.getRange());
 		main[2][2].setText("dps");
